// src/enum/error.enum.ts
export enum ErrorEnum {
  usernameRequired = 'Username is required',
  passwordRequired = 'Password is required',
  required = 'is required',
  userLockedOut = 'this user has been locked out.',
  firstNameRequired = 'First Name is required',
  lastNameRequired = 'Last Name is required',
  postalCodeRequired = 'Postal Code is required',
}