// src/enum/url.enum.ts
export enum PageURL {
  loginPageUrl = 'https://www.saucedemo.com',
  productPageUrl = 'https://www.saucedemo.com/inventory.html',
  cartPageUrl = 'https://www.saucedemo.com/cart.html',
  checkOutInformationPageUrl = 'https://www.saucedemo.com/checkout-step-one.html',
  checkOutCompletePageUrl = 'https://www.saucedemo.com/checkout-complete.html',
  checkOutOverViewPageUrl = 'https://www.saucedemo.com/checkout-step-two.html',
}